+++ tests/test_brats_format.py
@@ -41,4 +41,4 @@ class BratsImporterTest(TestCase):
 
         dataset = Dataset.import_from(DUMMY_DATASET_DIR, 'brats')
 
-        compare_datasets(self, expected_dataset, dataset, require_images=True)
+        compare_datasets(self, expected_dataset, dataset, require_media=True)